backend-api/pyproject.toml
[project]
name = "backend-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "en-core-web-sm",
    "fastapi>=0.123.0",
    "httpx>=0.28.1",
    "pdfplumber>=0.11.8",
    "pydantic>=2.12.5",
    "spacy>=3.8.11",
    "openai>=1.0.0",
    "uvicorn[standard]>=0.38.0",
    "python-docx>=1.2.0",
    "urllib3<2.0",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
